fix volume/sample rate min/max clamping

+++ src/defaults.hh
@@ -13,8 +13,8 @@ constexpr int UPDATE_RATE = 500; /* ui update rate (ms) */
 constexpr int IMAGE_UPDATE_RATE_LIMIT = 300; /* (ms) */
 constexpr int READ_TIMEOUT = 5000; /* string input timeout (ms) */
 constexpr f64 MPRIS_UPDATE_RATE = 100.0; /* delay between mpris polls (ms) */
-constexpr int MIN_SAMPLE_RATE = 1000;
-constexpr int MAX_SAMPLE_RATE = 9999999;
+constexpr u64 MIN_SAMPLE_RATE = 1000;
+constexpr u64 MAX_SAMPLE_RATE = 9999999;
 constexpr f64 FONT_ASPECT_RATIO = 1.0 / 2.0; /* typical monospaced font is 1/2 or 3/5 (width/height) */
 
 } /* namespace defaults */
